requirements-dev.txt
# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code quality
black>=23.0.0
ruff>=0.1.0
mypy>=1.5.0

# Testing utilities
httpx-mock>=0.2.0
pyyaml>=6.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.0.0